Crate external_storage

External storage support.

This crate define an abstraction of external storage. Currently, it supports local storage.

Modules

gcs

local

metrics

noop

s3

util

Structs

GCSStorage

LocalStorage
A storage saves files in local file system.

NoopStorage
A storage saves files into void. It is mainly for test use.

S3Storage
S3 compatible storage

Constants

READ_BUF_SIZE

Traits

ExternalStorage
An abstraction of an external storage.

Functions

block_on_external_io
Runs a future on the current thread involving external storage.

create_storage
Create a new storage from the given storage backend description.

make_gcs_backend

make_local_backend
Creates a local StorageBackend to the given path.

make_noop_backend
Creates a noop StorageBackend.

make_s3_backend

url_of_backend
Formats the storage backend as a URL.