Struct external_storage::s3::S3Uploader

struct S3Uploader<'client> {
    client: &'client S3Client,
    bucket: String,
    key: String,
    acl: Option<String>,
    server_side_encryption: Option<String>,
    ssekms_key_id: Option<String>,
    storage_class: Option<String>,
    upload_id: String,
    parts: Vec<CompletedPart>,
}

A helper for uploading a large files to S3 storage.

Note: this uploader does not support uploading files larger than 19.5 GiB.

Fields
client: &'client S3Clientbucket: Stringkey: Stringacl: Option<String>server_side_encryption: Option<String>ssekms_key_id: Option<String>storage_class: Option<String>upload_id: Stringparts: Vec<CompletedPart>
Implementations

impl<'client> S3Uploader<'client>

fn new(client: &'client S3Client, config: &Config, key: String) -> Self

Creates a new uploader with a given target location and upload configuration.

async fn run<'_>(
__arg0: Self,
reader: &'_ mut (dyn AsyncRead + Unpin),
est_len: u64
) -> Result<(), Box<dyn Error>>

Executes the upload process.

async fn begin<'_>(
&'_ self
) -> Result<String, RusotoError<CreateMultipartUploadError>>

Starts a multipart upload process.

async fn complete<'_>(
&'_ self
) -> Result<(), RusotoError<CompleteMultipartUploadError>>

Completes a multipart upload process, asking S3 to join all parts into a single file.

async fn abort<'_>(
&'_ self
) -> Result<(), RusotoError<AbortMultipartUploadError>>

Aborts the multipart upload process, deletes all uploaded parts.

async fn upload_part<'_, '_>(
&'_ self,
part_number: i64,
data: &'_ [u8]
) -> Result<CompletedPart, RusotoError<UploadPartError>>

Uploads a part of the file.

The part_number must be between 1 to 10000.

async fn upload<'_, '_>(
&'_ self,
data: &'_ [u8]
) -> Result<(), RusotoError<PutObjectError>>

Uploads a file atomically.

This should be used only when the data is known to be short, and thus relatively cheap to retry the entire upload.