Trait external_storage::ExternalStorage

pub trait ExternalStorage: 'static + Send + Sync {
    fn name(&self) -> &'static str;

    fn url(&self) -> Result<Url>;

    fn write(
        &self, 
        name: &str, 
        reader: Box<dyn AsyncRead + Send + Unpin>, 
        content_length: u64
    ) -> Result<()>;

    fn read(&self, name: &str) -> Box<dyn AsyncRead + Unpin>;

    fn restore(
        &self, 
        storage_name: &str, 
        restore_name: PathBuf, 
        expected_length: u64, 
        speed_limiter: &Limiter
    ) -> Result<()> { ... }
}

An abstraction of an external storage.

Required methods

fn name(&self) -> &'static str

fn url(&self) -> Result<Url>

fn write(
&self,
name: &str,
reader: Box<dyn AsyncRead + Send + Unpin>,
content_length: u64
) -> Result<()>

Write all contents of the read to the given path.

fn read(&self, name: &str) -> Box<dyn AsyncRead + Unpin>

Read all contents of the given path.

Loading content...
Provided methods

fn restore(
&self,
storage_name: &str,
restore_name: PathBuf,
expected_length: u64,
speed_limiter: &Limiter
) -> Result<()>

Read from external storage and restore to the given path

Loading content...